+++ pyo3-build-config/src/impl_.rs
@@ -100,6 +100,30 @@ pub struct InterpreterConfig {
     ///
     /// Serialized to `build_flags`.
     pub build_flags: BuildFlags,
+
+    /// Whether to suppress emitting of `cargo:rustc-link-*` lines from the build script.
+    ///
+    /// Typically, `pyo3`'s build script will emit `cargo:rustc-link-lib=` and
+    /// `cargo:rustc-link-search=` lines derived from other fields in this struct. In
+    /// advanced building configurations, the default logic to derive these lines may not
+    /// be sufficient. This field can be set to `Some(true)` to suppress the emission
+    /// of these lines.
+    ///
+    /// If suppression is enabled, `extra_build_script_lines` should contain equivalent
+    /// functionality or else a build failure is likely.
+    pub suppress_build_script_link_lines: bool,
+
+    /// Additional lines to `println!()` from Cargo build scripts.
+    ///
+    /// This field can be populated to enable the `pyo3` crate to emit additional lines from its
+    /// its Cargo build script.
+    ///
+    /// This crate doesn't populate this field itself. Rather, it is intended to be used with
+    /// externally provided config files to give them significant control over how the crate
+    /// is build/configured.
+    ///
+    /// Serialized to multiple `extra_build_script_line` values.
+    pub extra_build_script_lines: Vec<String>,
 }
 
 impl InterpreterConfig {
@@ -232,6 +256,8 @@ print("mingw", get_platform() == "mingw")
             executable: map.get("executable").cloned(),
             pointer_width: Some(calcsize_pointer * 8),
             build_flags: BuildFlags::from_interpreter(interpreter)?.fixup(version, implementation),
+            suppress_build_script_link_lines: false,
+            extra_build_script_lines: vec![],
         })
     }
 
@@ -271,6 +297,8 @@ print("mingw", get_platform() == "mingw")
         let mut executable = None;
         let mut pointer_width = None;
         let mut build_flags = None;
+        let mut suppress_build_script_link_lines = None;
+        let mut extra_build_script_lines = vec![];
 
         for (i, line) in lines.enumerate() {
             let line = line.context("failed to read line from config")?;
@@ -293,6 +321,12 @@ print("mingw", get_platform() == "mingw")
                 "executable" => parse_value!(executable, value),
                 "pointer_width" => parse_value!(pointer_width, value),
                 "build_flags" => parse_value!(build_flags, value),
+                "suppress_build_script_link_lines" => {
+                    parse_value!(suppress_build_script_link_lines, value)
+                }
+                "extra_build_script_line" => {
+                    extra_build_script_lines.push(value.to_string());
+                }
                 unknown => bail!("unknown config key `{}`", unknown),
             }
         }
@@ -318,6 +352,8 @@ print("mingw", get_platform() == "mingw")
                 }
                 .fixup(version, implementation)
             }),
+            suppress_build_script_link_lines: suppress_build_script_link_lines.unwrap_or(false),
+            extra_build_script_lines,
         })
     }
 
@@ -356,6 +392,11 @@ print("mingw", get_platform() == "mingw")
         write_option_line!(executable)?;
         write_option_line!(pointer_width)?;
         write_line!(build_flags)?;
+        write_line!(suppress_build_script_link_lines)?;
+        for line in &self.extra_build_script_lines {
+            writeln!(writer, "extra_build_script_line={}", line)
+                .context("failed to write extra_build_script_line")?;
+        }
         Ok(())
     }
 }
@@ -469,6 +510,7 @@ struct CrossCompileConfig {
     arch: String,
 }
 
+#[allow(unused)]
 pub fn any_cross_compiling_env_vars_set() -> bool {
     env::var_os("PYO3_CROSS").is_some()
         || env::var_os("PYO3_CROSS_LIB_DIR").is_some()
@@ -587,7 +629,7 @@ impl FromStr for BuildFlag {
 ///
 /// see Misc/SpecialBuilds.txt in the python source for what these mean.
 #[cfg_attr(test, derive(Debug, PartialEq))]
-#[derive(Default)]
+#[derive(Clone, Default)]
 pub struct BuildFlags(pub HashSet<BuildFlag>);
 
 impl BuildFlags {
@@ -928,6 +970,8 @@ fn load_cross_compile_from_sysconfigdata(
         executable: None,
         pointer_width,
         build_flags: BuildFlags::from_config_map(&sysconfig_data).fixup(version, implementation),
+        suppress_build_script_link_lines: false,
+        extra_build_script_lines: vec![],
     })
 }
 
@@ -947,6 +991,8 @@ fn windows_hardcoded_cross_compile(
         executable: None,
         pointer_width: None,
         build_flags: BuildFlags::windows_hardcoded(),
+        suppress_build_script_link_lines: false,
+        extra_build_script_lines: vec![],
     })
 }
 
@@ -1150,6 +1196,8 @@ mod tests {
             lib_dir: Some("lib_dir".into()),
             shared: true,
             version: MINIMUM_SUPPORTED_VERSION,
+            suppress_build_script_link_lines: true,
+            extra_build_script_lines: vec!["cargo:test1".to_string(), "cargo:test2".to_string()],
         };
         let mut buf: Vec<u8> = Vec::new();
         config.to_writer(&mut buf).unwrap();
@@ -1179,6 +1227,8 @@ mod tests {
                 major: 3,
                 minor: 10,
             },
+            suppress_build_script_link_lines: false,
+            extra_build_script_lines: vec![],
         };
         let mut buf: Vec<u8> = Vec::new();
         config.to_writer(&mut buf).unwrap();
@@ -1204,6 +1254,8 @@ mod tests {
                 executable: None,
                 pointer_width: None,
                 build_flags: BuildFlags::default(),
+                suppress_build_script_link_lines: false,
+                extra_build_script_lines: vec![],
             }
         )
     }
@@ -1313,7 +1365,9 @@ mod tests {
                 lib_dir: Some("C:\\some\\path".into()),
                 executable: None,
                 pointer_width: None,
-                build_flags: BuildFlags::windows_hardcoded()
+                build_flags: BuildFlags::windows_hardcoded(),
+                suppress_build_script_link_lines: false,
+                extra_build_script_lines: vec![],
             }
         );
     }
@@ -1379,6 +1433,8 @@ mod tests {
             lib_name: None,
             shared: true,
             version: PythonVersion { major: 3, minor: 7 },
+            suppress_build_script_link_lines: false,
+            extra_build_script_lines: vec![],
         };
 
         fixup_config_for_abi3(&mut config, Some(PythonVersion { major: 3, minor: 6 })).unwrap();
@@ -1397,6 +1453,8 @@ mod tests {
             lib_name: None,
             shared: true,
             version: PythonVersion { major: 3, minor: 6 },
+            suppress_build_script_link_lines: false,
+            extra_build_script_lines: vec![],
         };
 
         assert!(